Improve testing of catalog.Component
