+++ pkg/catalog/component/component.go
@@ -1,9 +1,9 @@
+// Package component provides high level structs and functions to create catalog entities
+// of type "component" in an opinionated way. It supports some well-known annotations and labels.
 package component
 
 import (
 	"fmt"
-	"sort"
-	"strings"
 	"time"
 
 	bscatalog "github.com/giantswarm/backstage-catalog-importer/pkg/bscatalog/v1alpha1"
@@ -92,6 +92,26 @@ type Component struct {
 	Spec bscatalog.ComponentSpec
 }
 
+func New(name string, options ...Option) (*Component, error) {
+	if name == "" {
+		return nil, fmt.Errorf("name must not be empty")
+	}
+
+	c := &Component{
+		Name:      name,
+		Namespace: "default",
+		Type:      "unspecified",
+		Owner:     "unspecified",
+		Lifecycle: "production",
+	}
+
+	for _, option := range options {
+		option(c)
+	}
+
+	return c, nil
+}
+
 // Add a tag to the Component
 func (c *Component) AddTag(tag string) {
 	c.Tags = append(c.Tags, tag)
@@ -118,107 +138,4 @@ func (c *Component) SetLabel(key, value string) {
 		c.Labels = make(map[string]string)
 	}
 	c.Labels[key] = value
-}
-
-// Returns an entity representation of the Component.
-func (c *Component) ToEntity() *bscatalog.Entity {
-	sort.Strings(c.Tags)
-
-	e := &bscatalog.Entity{
-		APIVersion: bscatalog.APIVersion,
-		Kind:       bscatalog.EntityKindComponent,
-		Metadata: bscatalog.EntityMetadata{
-			Annotations: make(map[string]string),
-			Description: c.Description,
-			Labels:      make(map[string]string),
-			Links:       make([]bscatalog.EntityLink, 0),
-			Name:        c.Name,
-			Title:       c.Title,
-		},
-	}
-
-	if c.Namespace != "default" {
-		e.Metadata.Namespace = c.Namespace
-	}
-
-	if c.Annotations != nil {
-		e.Metadata.Annotations = make(map[string]string)
-		for k, v := range c.Annotations {
-			e.Metadata.Annotations[k] = v
-		}
-	}
-
-	if c.Labels != nil {
-		e.Metadata.Labels = make(map[string]string)
-		for k, v := range c.Labels {
-			e.Metadata.Labels[k] = v
-		}
-	}
-
-	if c.Links != nil {
-		e.Metadata.Links = c.Links
-	}
-
-	if c.Tags != nil {
-		e.Metadata.Tags = c.Tags
-	}
-
-	if c.GithubProjectSlug != "" {
-		e.Metadata.Annotations["github.com/project-slug"] = c.GithubProjectSlug
-		e.Metadata.Annotations["backstage.io/source-location"] = fmt.Sprintf("url:https://github.com/%s", c.GithubProjectSlug)
-		if c.HasReadme && c.DefaultBranch != "" {
-			e.Metadata.Annotations["backstage.io/techdocs-ref"] = fmt.Sprintf("url:https://github.com/%s/tree/%s", c.GithubProjectSlug, c.DefaultBranch)
-		}
-	}
-	if c.QuayRepositorySlug != "" {
-		e.Metadata.Annotations["quay.io/repository-slug"] = c.QuayRepositorySlug
-	}
-	if c.GithubTeamSlug != "" {
-		e.Metadata.Annotations["github.com/team-slug"] = c.GithubTeamSlug
-	}
-	if c.OpsGenieTeam != "" {
-		e.Metadata.Annotations["opsgenie.com/team"] = c.OpsGenieTeam
-	}
-	if c.OpsGenieComponentSelector != "" {
-		e.Metadata.Annotations["opsgenie.com/component-selector"] = c.OpsGenieComponentSelector
-	}
-	if c.LatestReleaseTag != "" {
-		e.Metadata.Annotations["giantswarm.io/latest-release-tag"] = c.LatestReleaseTag
-	}
-	if c.LatestReleaseTime.Format(time.RFC3339) != "0001-01-01T00:00:00Z" {
-		e.Metadata.Annotations["giantswarm.io/latest-release-date"] = c.LatestReleaseTime.Format(time.RFC3339)
-	}
-	if c.CircleCiSlug != "" {
-		e.Metadata.Annotations["circleci.com/project-slug"] = c.CircleCiSlug
-	}
-	if c.DeploymentNames != nil {
-		sort.Strings(c.DeploymentNames)
-		e.Metadata.Annotations["giantswarm.io/deployment-names"] = strings.Join(c.DeploymentNames, ",")
-	}
-	if c.Type == "service" {
-		e.Metadata.Annotations["backstage.io/kubernetes-id"] = c.Name
-		if c.KubernetesID != "" {
-			e.Metadata.Annotations["backstage.io/kubernetes-id"] = c.KubernetesID
-		}
-	}
-
-	spec := bscatalog.ComponentSpec{
-		Type:      c.Type,
-		Lifecycle: c.Lifecycle,
-		Owner:     c.Owner,
-	}
-	if c.System != "" {
-		spec.System = c.System
-	}
-	if len(c.DependsOn) > 0 {
-		sort.Strings(c.DependsOn)
-		for i, d := range c.DependsOn {
-			c.DependsOn[i] = "component:" + d
-		}
-		spec.DependsOn = c.DependsOn
-	}
-
-	e.Spec = spec
-
-	return e
 }
